+++ app/Http/Controllers/GlossaryController.php
@@ -7,6 +7,7 @@ use Illuminate\Http\Request;
 use App\Http\Requests;
 
 use App\Models\Glossary;
+use App\Models\Material;
 
 class GlossaryController extends Controller
 {
@@ -27,6 +28,13 @@ class GlossaryController extends Controller
   }
 
   public function store(Request $request) {
+    $material = Material::find($request->input('material_id'));
+    $exercise = $material->exercise;
+    if(!Auth::user()->is_admin && $exercise->school == NULL)
+      return back()->withErrors('Et voi luoda tälle harjoitukselle sanastoa!');
+    if(!Auth::user()->is_admin && Auth::user()->school->id != $exercise->school->id)
+      return back()->withErrors('Et voi luoda tälle harjoitukselle sanatoa!');
+
     $glossary = new Glossary();
     $glossary->rus = $this->clearEmptyLines($request->input('rus'));
     $glossary->fin = $this->clearEmptyLines($request->input('fin'));
@@ -41,6 +49,12 @@ class GlossaryController extends Controller
 
   public function update(Request $request, $id) {
     $glossary = Glossary::find($id);
+    $exercise = $glossary->material->exercise;
+    if(!Auth::user()->is_admin && $exercise->school == NULL)
+      return back()->withErrors('Et voi päivittää tätä sanastoa!');
+    if(!Auth::user()->is_admin && Auth::user()->school->id != $exercise->school->id)
+      return back()->withErrors('Et voi päivittää tätä sanastoa!');
+
     $glossary->rus = $this->clearEmptyLines($request->input('rus'));
     $glossary->fin = $this->clearEmptyLines($request->input('fin'));
     if ($this->checkLines($glossary->rus,$glossary->fin)) {
@@ -52,6 +66,13 @@ class GlossaryController extends Controller
   }
 
   public function destroy($id) {
+    $glossary = Glossary::find($id);
+    $exercise = $glossary->material->exercise;
+    if(!Auth::user()->is_admin && $exercise->school == NULL)
+      return back()->withErrors('Et voi poistaa tätä sanastoa!');
+    if(!Auth::user()->is_admin && Auth::user()->school->id != $exercise->school->id)
+      return back()->withErrors('Et voi poistaa tätä sanastoa!');
+
     $glossary = Glossary::find($id);
     $glossary->delete();
     return back()->with('success', 'Sanasto poistettu!');
